+++ Ambev.DeveloperEvaluation/backend/src/Ambev.DeveloperEvaluation.Domain/Entities/Sale.cs
@@ -5,7 +5,7 @@ namespace Ambev.DeveloperEvaluation.Domain.Entities;
 
 public class Sale : BaseEntity
 {
-    public int Number { get; set; }
+    public Int64 Number { get; set; }
 
     public DateTime Date { get; set; }
